added login decorator to 'my products', moved all product list by type functionality to a new view called product_type_view and made necessary alterations to init and views.py.
