+++ website/views/__init__.py
@@ -1,6 +1,7 @@
-from .views import index, register, login_user, user_logout, list_products, products_by_type
+from .views import index, register, login_user, user_logout, list_products
 
 from .profile_list_view import *
 from .products_home_view import productHomeView 
 from .product_category_view import sell_product
 from .product_detail_views import product_details
+from .product_type_view import products_by_type
