+++ website/views/views.py
@@ -5,7 +5,7 @@ from django.shortcuts import render
 from django.template import RequestContext
 
 from website.forms import UserForm, ProductForm
-from website.models import Product, ProductType
+from website.models import Product
 
 def index(request):
     template_name = 'index.html'
@@ -92,36 +92,13 @@ def user_logout(request):
     # in the URL in redirects?????
     return HttpResponseRedirect('/')
 
+@login_required
 def list_products(request):
     user_id = request.user.id
     all_products = Product.objects.raw('''select * from website_product
 where customer_id = %s ''', [user_id] )
     template_name = 'product/list.html'
     return render(request, template_name, {'products': all_products})
-
-
-def products_by_type(request, pk):
-    '''
-    This method gets one individual producttype and lists it along with all associated products by name, quantity and price.
-
-    Method uses a SELECT from the product table and JOIN with the producttype table.
-    '''
-    try:
-
-        sql = '''
-        SELECT p.title, p.quantity, p.price, pt.name, pt.id, p.id
-        FROM website_product p
-        JOIN website_producttype pt on pt.id = p.productType_id
-        WHERE pt.id=%s
-        '''
-        product_type = ProductType.objects.raw(sql, [pk])[0]
-        product = Product.objects.raw(sql, [pk])
-
-    except IndexError:
-        raise Http404("This product type does not exist")
-
-    context = {'type': product_type, 'prod':product}
-    return render(request, 'product/products_by_type.html', context)
